requirements.txt
aniso8601==8.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
flask-unittest==0.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Padding==0.5
pycryptodome==3.10.1
pytz==2021.1
requests==2.25.1
six==1.15.0
testcase==0.1.0
urllib3==1.26.3
Werkzeug==1.0.1
